+++ src/Component/Feed/FeedCard/CardFeed/CardFeed.jsx
@@ -1,23 +1,15 @@
-import React, {useState} from 'react';
+import React from 'react';
 import PropTypes from "prop-types";
 // css
 import {
-    CardContent,
-    CardFeedDesc,
     CardFeedTitle,
-    CardFeedWrapper, UnfoldBtn,
+    CardFeedWrapper,
 } from "./CardFeed.styles";
 // Component
 import CardImage from "./CardImage";
+import CardContent from "./CardContent";
 
 const CardFeed = ({title, feedContent, fileNames}) => {
-    const [open, setOpen] = useState(false);
-
-    const openContent = (e) => {
-        e.preventDefault();
-        setOpen(true);
-    }
-
     return (
         <CardFeedWrapper>
             <CardFeedTitle>
@@ -25,19 +17,7 @@ const CardFeed = ({title, feedContent, fileNames}) => {
                     {title}
                 </h3>
             </CardFeedTitle>
-            <CardFeedDesc>
-                <CardContent dangerouslySetInnerHTML={{
-                    __html: feedContent
-                }}
-                             open={open}
-                />
-                <UnfoldBtn type={"button"}
-                           onClick={openContent}
-                           open={open}
-                >
-                    ... 더보기
-                </UnfoldBtn>
-            </CardFeedDesc>
+            <CardContent feedContent={feedContent}/>
             {
                 fileNames.length > 0 ?
                     <CardImage fileNames={fileNames}/>
